Fixed minor things

+++ data.h
@@ -5,7 +5,7 @@ typedef struct DataStruct *Data;
 typedef const struct DataStruct *ConstData;
 
 Data new_data(unsigned char minuti, unsigned char ora, unsigned char giorno,
-              unsigned char mese, unsigned int anno);
+              unsigned char mese, unsigned short anno);
 int is_valid_data(ConstData data);
 int is_leap(ConstData data);
 int cmp_data(ConstData data_a, ConstData data_b);
